[rpc] Move error checks to blockchain
