+++ internal/hmyapi/apiv1/error.go
@@ -2,9 +2,6 @@ package apiv1
 
 import (
 	"errors"
-
-	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/harmony-one/harmony/shard"
 )
 
 var (
@@ -19,17 +16,3 @@ var (
 	// ErrRequestedBlockTooHigh when given block is greater than latest block number
 	ErrRequestedBlockTooHigh = errors.New("requested block number greater than current block number")
 )
-
-func (s *PublicBlockChainAPI) isBeaconShard() error {
-	if s.b.GetShardID() != shard.BeaconChainShardID {
-		return ErrNotBeaconShard
-	}
-	return nil
-}
-
-func (s *PublicBlockChainAPI) isBlockGreaterThanLatest(blockNum rpc.BlockNumber) error {
-	if uint64(blockNum) > s.b.CurrentBlock().NumberU64() {
-		return ErrRequestedBlockTooHigh
-	}
-	return nil
-}
